refactor: lift up `handleUriCall` to server
