+++ packages/serverpod/lib/src/server/endpoint_dispatch.dart
@@ -1,4 +1,3 @@
-import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:serverpod/src/authentication/authentication_info.dart';
@@ -188,117 +187,6 @@ abstract class EndpointDispatch {
   String _endpointFromName(String name) {
     var components = name.split('/');
     return components[0];
-  }
-
-  /// Dispatches a call to the [Server] to the correct [Endpoint] method. If
-  /// successful, it returns the object from the method. If unsuccessful it will
-  /// return a [Result] object.
-  Future<Result> handleUriCall(
-    Server server,
-    String path,
-    Uri uri,
-    String body,
-    HttpRequest request,
-  ) async {
-    var endpointComponents = path.split('.');
-    if (endpointComponents.isEmpty || endpointComponents.length > 2) {
-      return ResultInvalidParams('Endpoint $path is not a valid endpoint name');
-    }
-
-    // Read query parameters
-    var isValidBody = body != '' && body != 'null';
-    Map<String, dynamic> queryParameters =
-        isValidBody ? decodeParameters(body) : {};
-
-    // Add query parameters from uri
-    queryParameters.addAll(uri.queryParameters);
-
-    String endpointName;
-    String methodName;
-
-    if (path.contains('/')) {
-      // Using the new path format (for OpenAPI)
-      var pathComponents = path.split('/');
-      endpointName = pathComponents[0];
-      methodName = pathComponents[1];
-    } else {
-      // Using the standard format with query parameters
-      endpointName = path;
-      var method = queryParameters['method'];
-      if (method is String) {
-        methodName = method;
-      } else {
-        return ResultInvalidParams(
-          'No method name specified in call to $endpointName',
-        );
-      }
-    }
-
-    // Get the the authentication key, if any
-    String? authenticationKey = queryParameters['auth'];
-
-    late MethodCallSession session;
-
-    Endpoint endpoint;
-    MethodConnector method;
-    Map<String, dynamic> paramMap;
-    try {
-      (endpoint, method, paramMap) = await getAuthorizedEndpointMethodConnector(
-        createSessionCallback: (connector) {
-          session = MethodCallSession(
-            server: server,
-            uri: uri,
-            body: body,
-            path: path,
-            httpRequest: request,
-            method: methodName,
-            endpoint: endpointName,
-            queryParameters: queryParameters,
-            authenticationKey: authenticationKey,
-            enableLogging: connector.endpoint.logSessions,
-          );
-          return session;
-        },
-        endpointPath: endpointName,
-        methodName: methodName,
-        parameters: queryParameters,
-        serializationManager: server.serializationManager,
-      );
-    } on MethodNotFoundException catch (e) {
-      return ResultInvalidParams(e.message);
-    } on InvalidEndpointMethodTypeException catch (e) {
-      return ResultInvalidParams(e.message);
-    } on EndpointNotFoundException catch (e) {
-      return ResultInvalidParams(e.message);
-    } on NotAuthorizedException catch (e) {
-      return e.authenticationFailedResult;
-    } on InvalidParametersException catch (e, stackTrace) {
-      var sessionLogId = await session.close(error: e, stackTrace: stackTrace);
-      return ResultInternalServerError(
-          e.toString(), stackTrace, sessionLogId ?? 0);
-    }
-
-    try {
-      var result = await method.call(session, paramMap);
-
-      return ResultSuccess(
-        result,
-        sendByteDataAsRaw: endpoint.sendByteDataAsRaw,
-      );
-    } on SerializableException catch (exception) {
-      return ExceptionResult(model: exception);
-    } on Exception catch (e, stackTrace) {
-      var sessionLogId = await session.close(error: e, stackTrace: stackTrace);
-      return ResultInternalServerError(
-          e.toString(), stackTrace, sessionLogId ?? 0);
-    } catch (e, stackTrace) {
-      // Something did not work out
-      var sessionLogId = await session.close(error: e, stackTrace: stackTrace);
-      return ResultInternalServerError(
-          e.toString(), stackTrace, sessionLogId ?? 0);
-    } finally {
-      await session.close();
-    }
   }
 
   /// Checks if a user can access an [Endpoint]. If access is granted null is
